+++ packages/at_persistence_secondary_server/lib/at_persistence_secondary_server.dart
@@ -2,6 +2,7 @@ export 'package:at_persistence_secondary_server/src/factory/at_persistence_confi
 export 'package:at_persistence_secondary_server/src/factory/at_persistence_factory.dart';
 export 'package:at_persistence_secondary_server/src/factory/hive_at_persistence_factory.dart';
 export 'package:at_persistence_secondary_server/src/compaction/at_compaction_job.dart';
+export 'package:at_persistence_secondary_server/src/compaction/at_compaction_strategy.dart';
 export 'package:at_persistence_secondary_server/src/compaction/commit_log_compaction_service.dart';
 export 'package:at_persistence_secondary_server/src/compaction/at_compaction_stats_service.dart';
 export 'package:at_persistence_secondary_server/src/compaction/at_compaction_stats_service_impl.dart';